Better Time Sync and Error checking
As it takes about 8 thousand cycles to load the instrument program into a cog, I decided to load the parameter values AFTER launching the cog, as there is time to do this before the cog even starts. This additionally gives me time to do some sanity checks of the values as well (with no extra cycle costs).

Additionally I wait until the cog is completely launched before doing any more instructions. This is from the cog setting the first global parameter to all 1's (after the real value is retrieved of course). SPIN knows to wait for this value before continuing